Drop utils.go for both implementations.
